requirements.txt
Django>=4.0,<5.0
djangorestframework
psycopg2-binary
drf-yasg
flake8